Refactor and restructure HRL training environment
- Removed `train_with_cus_policy.py` as it contained commented-out code for SAC and PPO models.
- Updated `worker_env.py` to enhance the `TableTennisWorker` class with improved goal sampling and reward computation.
- Deleted `manager_env.py` as it was redundant and replaced by a more integrated approach in `train_ppo_hrl.py`.
- Introduced `train_ppo_hrl.py` to manage the training process for both worker and manager policies, including callbacks for evaluation and logging.
- Adjusted `train_ppo.py` and `train_ppo_lattice.py` to accommodate changes in the video callback initialization.
- Removed deprecated code and ensured compatibility with the new environment structure.
