+++ train_ppo.py
@@ -46,7 +46,12 @@ def main():
         render=False,
     )
     
-    video_cb = VideoCallback(CustomEnv, cfg, make_predict_fn(model))
+    video_cb = VideoCallback(
+        env_func=CustomEnv,
+        env_args={"config": cfg},
+        cfg=cfg,
+        predict_fn=make_predict_fn(model)
+    )
     
     model.learn(
         total_timesteps=cfg.ppo_total_timesteps,
