+++ train_ppo_hrl.py
@@ -2,7 +2,6 @@ import copy
 import os
 from stable_baselines3.common.callbacks import CallbackList, EvalCallback
 from stable_baselines3 import PPO
-from lattice.ppo.policies import LatticeRecurrentActorCriticPolicy
 
 from config import Config
 from env_factory import build_env
@@ -10,8 +9,8 @@ from utils import prepare_experiment_directory, make_predict_fn
 from callbacks.infologger_callback import InfoLoggerCallback
 from callbacks.video_callback import VideoCallback
 
-from worker_env import TableTennisWorker
-from manager_env import TableTennisManager
+from hrl.worker_env import TableTennisWorker
+from hrl.manager_env import TableTennisManager
 from dr_spcrl.dr_spcrl import DRSPCRLRecurrentPPO
 from loguru import logger
 
@@ -25,35 +24,22 @@ def main():
     worker_cfg.logdir = os.path.join(cfg.logdir, "worker")
     env_worker = build_env(worker_cfg, env_type="worker")
 
-    worker_model = DRSPCRLRecurrentPPO(
-        policy=LatticeRecurrentActorCriticPolicy,
-        env=env_worker,
-        tensorboard_log=worker_cfg.logdir,
+    worker_model = PPO(
+        "MlpPolicy",
+        env_worker,
         verbose=1,
-        device="auto",
-        batch_size=worker_cfg.ppo_batch_size,
-        n_steps=worker_cfg.ppo_n_steps,
-        n_epochs=worker_cfg.ppo_epochs,
-        learning_rate=worker_cfg.ppo_lr,
-        clip_range=worker_cfg.ppo_clip,
-        gamma=worker_cfg.ppo_gamma,
-        gae_lambda=worker_cfg.ppo_lambda,
-        ent_coef=3.62109e-06,
-        max_grad_norm=0.7,
-        vf_coef=0.835671,
-        policy_kwargs=dict(
-            use_lattice=True,
-            use_expln=True,
-            ortho_init=False,
-            log_std_init=0.0,
-            std_clip=(1e-3, 10),
-        ),
-        eps_start=0.0,
-        eps_budget=1.0,
-        lr_beta=5e-4,
-        beta_updates=5,
-        lr_curr=1e-3,
-        alpha=0.1,
+        tensorboard_log=os.path.join(cfg.logdir),
+        n_steps=cfg.ppo_n_steps,
+        batch_size=cfg.ppo_n_steps//cfg.ppo_batch_size,
+        gamma=cfg.ppo_gamma,
+        learning_rate=cfg.ppo_lr,
+        gae_lambda=cfg.ppo_lambda,
+        n_epochs=cfg.ppo_epochs,
+        max_grad_norm=cfg.ppo_max_grad_norm,
+        clip_range=cfg.ppo_clip_range,
+        seed=cfg.seed,
+        clip_range_vf=cfg.ppo_clip_range,
+        policy_kwargs=dict(net_arch=[cfg.ppo_hidden_dim]),
     )
 
     # # Callback Share
@@ -72,8 +58,7 @@ def main():
         deterministic=True,
         render=False,
     )
-    video_worker_env = TableTennisWorker(worker_cfg)
-    video_cb = VideoCallback(video_worker_env, worker_cfg, make_predict_fn(worker_model))
+    video_cb = VideoCallback(TableTennisWorker, worker_cfg, make_predict_fn(worker_model))
 
     # Learn
     logger.info("🚧 Training Worker Policy 🚧")
@@ -133,13 +118,17 @@ def main():
         render=False,
     )
     video_worker_env = TableTennisWorker(worker_cfg)
-    video_manager_env = TableTennisManager(
-        worker_env=video_worker_env,
-        worker_model=worker_model,
-        config=manager_cfg,
+    video_cb = VideoCallback(
+        env_func=TableTennisManager,
+        env_args={
+            "worker_env": video_worker_env,
+            "worker_model": worker_model,
+            "config": manager_cfg,
+        },
+        cfg=manager_cfg,
+        predict_fn=make_predict_fn(manager_model)
     )
-    video_cb = VideoCallback(video_manager_env, manager_cfg, make_predict_fn(manager_model))
-    
+        
     # Learn
     logger.info("🚧 Training Manager Policy 🚧")
     manager_model.learn(
@@ -151,7 +140,6 @@ def main():
     
     eval_manager_env.close()
     video_worker_env.close()
-    video_manager_env.close()
     env_manager.close()
     
     logger.info("🏁 Training Complete 🏁")
